Almost Compiling

+++ linerow.h
@@ -20,7 +20,7 @@ typedef line_t* line_p;
 class LineRow
 {
 public:
-  LineRow(int size,int density,int maxDifference);
+  LineRow(int size, int density, float maxDifference);
   void setNewLine(pcl::PointCloud<pcl::PointXYZ> cloudPiece, int position);
   line_p getRowArray();
 };
